yunwei2.0.8/requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.1.0
requests>=2.31.0
paramiko>=3.3.1
prometheus-client>=0.18.0
python-dotenv>=1.0.0
asyncio>=3.4.3
typing-extensions>=4.8.0
pydantic>=2.5.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
python-multipart>=0.0.6
pymysql>=1.1.0